move OmicronPhysicalDiskConfig
