+++ sled-agent/config-reconciler/src/ledger.rs
@@ -682,7 +682,6 @@ mod tests {
     use iddqd::IdOrdMap;
     use illumos_utils::zpool::ZpoolName;
     use omicron_common::disk::DiskIdentity;
-    use omicron_common::disk::OmicronPhysicalDiskConfig;
     use omicron_test_utils::dev;
     use omicron_test_utils::dev::poll::wait_for_watch_channel_condition;
     use omicron_uuid_kinds::InternalZpoolUuid;
@@ -690,6 +689,7 @@ mod tests {
     use omicron_uuid_kinds::OmicronZoneUuid;
     use omicron_uuid_kinds::PhysicalDiskUuid;
     use omicron_uuid_kinds::ZpoolUuid;
+    use sled_agent_types::disk::OmicronPhysicalDiskConfig;
     use sled_agent_types::inventory::HostPhase2DesiredContents;
     use sled_agent_types::inventory::HostPhase2DesiredSlots;
     use sled_agent_types::inventory::OmicronZoneConfig;
