+++ sled-agent/config-reconciler/src/reconciler_task/external_disks.rs
@@ -22,11 +22,11 @@ use illumos_utils::zpool::Zpool;
 use illumos_utils::zpool::ZpoolName;
 use key_manager::StorageKeyRequester;
 use omicron_common::api::external::ByteCount;
-use omicron_common::disk::OmicronPhysicalDiskConfig;
 use omicron_uuid_kinds::PhysicalDiskUuid;
 use omicron_uuid_kinds::ZpoolUuid;
 use rand::distr::{Alphanumeric, SampleString};
 use sled_agent_types::disk::DiskVariant;
+use sled_agent_types::disk::OmicronPhysicalDiskConfig;
 use sled_agent_types::inventory::ConfigReconcilerInventoryResult;
 use sled_agent_types::inventory::ZpoolHealth;
 use sled_storage::config::MountConfig;
