+++ sled-agent/types/versions/src/initial/inventory.rs
@@ -20,9 +20,7 @@ use omicron_common::api::external::Generation;
 use omicron_common::api::external::MacAddr;
 use omicron_common::api::external::Name;
 use omicron_common::api::external::Vni;
-use omicron_common::disk::{
-    DatasetConfig, DatasetName, OmicronPhysicalDiskConfig,
-};
+use omicron_common::disk::{DatasetConfig, DatasetName};
 use omicron_common::snake_case_result;
 use omicron_common::snake_case_result::SnakeCaseResult;
 use omicron_common::update::OmicronInstallManifestSource;
@@ -44,6 +42,7 @@ use uuid::Uuid;
 
 use super::disk::DiskVariant;
 use super::disk::M2Slot;
+use super::disk::OmicronPhysicalDiskConfig;
 use crate::impls::inventory::SourceNatConfigError;
 
 /// Describes properties that should uniquely identify a Gimlet.
