+++ sled-agent/types/versions/src/measurements/inventory.rs
@@ -13,7 +13,7 @@ use iddqd::id_upcast;
 use omicron_common::api::external;
 use omicron_common::{
     api::external::{ByteCount, Generation},
-    disk::{DatasetConfig, OmicronPhysicalDiskConfig},
+    disk::DatasetConfig,
 };
 use omicron_ledger::Ledgerable;
 use omicron_uuid_kinds::SledUuid;
@@ -25,6 +25,7 @@ use sled_hardware_types::SledCpuFamily;
 use std::time::Duration;
 
 use crate::v1;
+use crate::v1::disk::OmicronPhysicalDiskConfig;
 use crate::v1::inventory::{
     BootPartitionContents, ConfigReconcilerInventoryResult,
     HostPhase2DesiredSlots, InventoryDataset, InventoryDisk, InventoryZpool,
